2022.02.04

+++ Practice/BaekJoon/9461/Main.java
@@ -4,7 +4,7 @@ import java.io.*;
 import java.util.*;
 
 public class Main {
-	public static int pn[] = new int[101];
+	public static long pn[] = new long[101];
 	
 	public static void main(String args[]) throws Exception {
 		BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
@@ -34,7 +34,7 @@ public class Main {
 		br.close();
 	}
 	
-	public static int PN(int input) {
+	public static long PN(int input) {
 		if(input > 2 && pn[input] == 0) {
 			pn[input] = PN(input-2) + PN(input-3);
 		}
